# Full ASM AVX2 accelerated version of the SAD functions